requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
swagger-ui-py==22.7.13